Configure production lobby to not auto-restart to pick up latest changes
Only have the prerelease lobby restart on deployments. Production
lobby we are okay getting a new version deployed but restarts are better
triggered by hand so we can time them better.

As part of this change we are also making the systemd reload be smarter
and only doing that if the service script has changed.

Note: 'using_latest' is true for prerelease which means we are always
deploying the latest jar file. That will in effect be on every merge,
every merge creates a new jar file and will trigger deployment.
Hence, detecting if a jar file changed and if using latest is very
nearly equivalent.