Convert GameParser to use new 'xml-reader' (StAX based parser)
GameParser instead of directly traversing an XML DOM instead now loads
a java-model of the XML and then traverses that model to create
a 'GameData'. The process of converting this java model to a 'GameData'
is referred to as 'GameData assembly'.

This update relaxes some game parsing validation and future updates
will need to re-work in map-data integrity checks back into parsing.